+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Modified
 * Avoid installation problems with setuptools==50
 * Migrating codebase to Python 3.6+ specific features (Lexicon will explicitly break on older versions now)
+* Fix Easyname provider to work with their new website
 
 ## 3.4.5 - 02/11/2020
 ### Added